pyproject.toml
[tool.poetry]
name = "dog_finder"
version = "0.1.0"
description = ""
authors = ["lilimik <kazakovlim@inbox.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
factory-boy = "^3.2.0"
psycopg2-binary = "^2.8.6"
django-choices = "^1.7.1"
Pillow = "^8.2.0"
python-dotenv = "^0.17.1"
djangorestframework-simplejwt = "^4.7.1"
drf-yasg = "^1.20.0"
djangorestframework = "^3.12.4"
Django = "~3.1"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
